Overall heavy code cleanup and performance imporvement.
Fixed, setStation still going on, even if the name was to long, same for
setDestination. Made iteration on checkStation much faster. Made
getStations and getDestinations faster. Cleaned code, so the sourcode
looks a lot better. Added various missing comments in the sourcecode.
Still a lot of code to move/cleaning up but this is the start.